Move DarkModeContext into components/utils
