+++ site/source/pages/Landing/SearchOrCreate.tsx
@@ -9,8 +9,8 @@ import {
 } from '@/api/fabrique-social'
 import { CompanyDetails } from '@/components/company/Details'
 import { CompanySearchField } from '@/components/company/SearchField'
+import { ForceThemeProvider } from '@/components/utils/DarkModeContext'
 import { useEngine } from '@/components/utils/EngineContext'
-import { ForceThemeProvider } from '@/contexts/DarkModeContext'
 import AnswerGroup from '@/design-system/answer-group'
 import { Button } from '@/design-system/buttons'
 import { Emoji } from '@/design-system/emoji'
