+++ site/source/pages/gerer/index.tsx
@@ -30,12 +30,12 @@ import { PlaceDesEntreprisesButton } from '@/components/PlaceDesEntreprises'
 import { CompanyDetails } from '@/components/company/Details'
 import RuleInput from '@/components/conversation/RuleInput'
 import { FromTop } from '@/components/ui/animate'
+import { ForceThemeProvider } from '@/components/utils/DarkModeContext'
 import DefaultHelmet from '@/components/utils/DefaultHelmet'
 import { useEngine } from '@/components/utils/EngineContext'
 import { ScrollToTop } from '@/components/utils/Scroll'
 import { Markdown } from '@/components/utils/markdown'
 import useSimulationConfig from '@/components/utils/useSimulationConfig'
-import { ForceThemeProvider } from '@/contexts/DarkModeContext'
 import { Message, Popover } from '@/design-system'
 import { Button } from '@/design-system/buttons'
 import { Emoji } from '@/design-system/emoji'
